breaking(core): Migrate module to Typescript and change build and test workflows
