+++ package.json
@@ -4,9 +4,9 @@
   "scripts": {
     "prepare": "lerna bootstrap --no-ci && lerna run build",
     "build": "lerna run build",
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
+    "test": "lerna run test",
+    "test:coverage": "lerna run test:coverage",
+    "coveralls": "lerna run coveralls",
     "lint": "eslint '**/*.js'",
     "prettier": "prettier '**/*.js' '**/*.md' '**/*.ts' --write",
     "release": "lerna publish",
@@ -27,6 +27,10 @@
     "@testing-library/jest-dom": "^5.1.1",
     "@testing-library/react": "^10.0.1",
     "@testing-library/react-hooks": "^3.2.1",
+    "@types/jest": "^26.0.5",
+    "@types/react": "^16.9.43",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-test-renderer": "^16.9.2",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
@@ -40,6 +44,7 @@
     "jest": "^25.1.0",
     "lerna": "^3.20.2",
     "lint-staged": "^10.0.8",
+    "npm-run-all": "^4.1.5",
     "prettier": "^2.0.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -49,7 +54,10 @@
     "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-peer-deps-external": "^2.2.2",
     "rollup-plugin-size-snapshot": "^0.11.0",
-    "rollup-plugin-terser": "^5.3.0"
+    "rollup-plugin-terser": "^5.3.0",
+    "ts-jest": "^26.1.3",
+    "tweak-sourcemap-paths": "^0.0.4",
+    "typescript": "^3.9.7"
   },
   "lint-staged": {
     "**/*.js": [
