+++ packages/core/.gitignore
@@ -1,0 +1,2 @@
+# Output folders
+lib*/
