--- packages/core/.size-snapshot.json
@@ -1,21 +1,0 @@
-{
-  "dist/lib/react-keycloak-core.js": {
-    "bundled": 7652,
-    "minified": 4082,
-    "gzipped": 1301
-  },
-  "dist/es/react-keycloak-core.js": {
-    "bundled": 7294,
-    "minified": 3793,
-    "gzipped": 1228,
-    "treeshaked": {
-      "rollup": {
-        "code": 249,
-        "import_statements": 249
-      },
-      "webpack": {
-        "code": 1431
-      }
-    }
-  }
-}
