+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@react-keycloak/core",
-  "version": "2.2.1",
+  "version": "3.0.0-alpha.2",
   "license": "MIT",
   "description": "React bindings for Keycloak javascript adapter (core)",
   "repository": {
@@ -14,15 +14,15 @@
   "homepage": "https://github.com/react-keycloak/react-keycloak#readme",
   "keywords": [
     "react",
-    "keycloak",
-    "keycloak-js"
+    "keycloak"
   ],
-  "main": "dist/lib/react-keycloak-core.js",
-  "module": "dist/es/react-keycloak-core.js",
-  "types": "index.d.ts",
+  "main": "lib-commonjs/index.js",
+  "jsnext:main": "lib/index.js",
+  "module": "lib/index.js",
+  "typings": "lib/index.d.ts",
   "files": [
-    "dist",
-    "index.d.ts"
+    "lib",
+    "lib-commonjs"
   ],
   "publishConfig": {
     "access": "public"
@@ -32,18 +32,27 @@
     "url": "https://www.patreon.com/reactkeycloak"
   },
   "scripts": {
-    "clean": "rimraf ./dist",
+    "clean": "rimraf lib lib-commonjs",
     "prebuild": "npm run clean",
-    "build": "rollup -c",
-    "prepublishOnly": "npm run build && cp ../../README.md . && cp ../../LICENSE.md ."
+    "compile:commonjs": "tsc --module commonjs --outDir lib-commonjs",
+    "compile:esnext": "tsc --module esnext --outDir lib",
+    "compile": "run-p compile:*",
+    "fixmaps:lib": "tweak_sourcemap_paths -a --include 'lib/**/*.js.map' 'lib-commonjs/**/*.js.map'",
+    "fixmaps": "run-p fixmaps:*",
+    "build": "run-s compile fixmaps",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
+    "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
+    "prepublishOnly": "yarn build && cp ../../README.md . && cp ../../LICENSE.md ."
   },
   "peerDependencies": {
-    "keycloak-js": ">=9.0.2",
     "react": ">=16"
   },
   "dependencies": {
-    "@babel/runtime": "^7.9.0",
-    "prop-types": "^15.7.2",
-    "react-fast-compare": "^3.0.1"
+    "react-fast-compare": "^3.2.0"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "jsdom"
   }
 }
