--- packages/core/src/context.js
@@ -1,10 +1,0 @@
-import { createContext } from 'react'
-
-export function createReactKeycloakContext(initialContext) {
-  return createContext({
-    initialized: false,
-    ...initialContext,
-  })
-}
-
-export default createReactKeycloakContext
