+++ packages/core/src/index.ts
@@ -2,3 +2,4 @@
 
 export * from './context'
 export * from './provider'
+export * from './types'
