--- packages/core/test/setupAfterEnv.js
@@ -1,1 +1,0 @@
-afterEach(require('@testing-library/react').cleanup)
